Move spring.http.* config properties namespace
Closes gh-18827
