+++ spring-boot-project/spring-boot-autoconfigure/src/test/java/org/springframework/boot/autoconfigure/http/codec/CodecsAutoConfigurationTests.java
@@ -1,5 +1,5 @@
 /*
- * Copyright 2012-2019 the original author or authors.
+ * Copyright 2012-2020 the original author or authors.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -23,7 +23,6 @@ import org.junit.jupiter.api.Test;
 
 import org.springframework.boot.autoconfigure.AutoConfigurations;
 import org.springframework.boot.autoconfigure.codec.CodecProperties;
-import org.springframework.boot.autoconfigure.http.HttpProperties;
 import org.springframework.boot.test.context.runner.WebApplicationContextRunner;
 import org.springframework.boot.web.codec.CodecCustomizer;
 import org.springframework.context.annotation.Bean;
@@ -59,7 +58,7 @@ class CodecsAutoConfigurationTests {
 
 	@Test
 	void loggingRequestDetailsCustomizerShouldUseHttpProperties() {
-		this.contextRunner.withPropertyValues("spring.http.log-request-details=true").run((context) -> {
+		this.contextRunner.withPropertyValues("spring.codec.log-request-details=true").run((context) -> {
 			CodecCustomizer customizer = context.getBean(CodecCustomizer.class);
 			CodecConfigurer configurer = new DefaultClientCodecConfigurer();
 			customizer.customize(configurer);
@@ -72,7 +71,7 @@ class CodecsAutoConfigurationTests {
 		this.contextRunner.run((context) -> {
 			Method customizerMethod = ReflectionUtils.findMethod(
 					CodecsAutoConfiguration.DefaultCodecsConfiguration.class, "defaultCodecCustomizer",
-					HttpProperties.class, CodecProperties.class);
+					CodecProperties.class);
 			Integer order = new TestAnnotationAwareOrderComparator().findOrder(customizerMethod);
 			assertThat(order).isEqualTo(0);
 		});
