+++ spring-boot-project/spring-boot-docs/src/docs/asciidoc/using-spring-boot.adoc
@@ -689,7 +689,7 @@ Rather than needing to set these properties manually, the `spring-boot-devtools`
 
 Because you need more information about web requests while developing Spring MVC and Spring WebFlux applications, developer tools will enable `DEBUG` logging for the `web` logging group.
 This will give you information about the incoming request, which handler is processing it, the response outcome, etc.
-If you wish to log all request details (including potentially sensitive information), you can turn on the configprop:spring.http.log-request-details[] configuration property.
+If you wish to log all request details (including potentially sensitive information), you can turn on the configprop:spring.mvc.log-request-details[] or configprop:spring.codec.log-request-details[] configuration properties.
 
 NOTE: If you don't want property defaults to be applied you can set configprop:spring.devtools.add-properties[] to `false` in your `application.properties`.
